+++ bin.js
@@ -3,8 +3,8 @@
 import { resolve } from 'path'
 import { existsSync } from './src/fs.js'
 
-if (existsSync('./node_modules/lazyrepo/src/index.js')) {
-  await import(resolve('./node_modules/lazyrepo/src/index.js'))
+if (existsSync('./node_modules/lazyrepo/src/cli.js')) {
+  await import(resolve('./node_modules/lazyrepo/src/cli.js'))
 } else {
-  await import('./src/index.js')
+  await import('./src/cli.js')
 }
