--- src/index.js
@@ -1,3 +1,0 @@
-import { exec } from './cli.js'
-
-await exec(process.argv)
